+++ SerialPort/serialport.h
@@ -9,8 +9,8 @@ public:
 	
 	virtual bool lock() = 0;
 	virtual void unlock() = 0;
-	virtual uint16_t read(void *buf, uint16_t size) = 0;
-	virtual void write(const void *buf, uint16_t size) = 0;
+	virtual size_t read(void *buf, size_t size) = 0;
+	virtual void write(const void *buf, size_t size) = 0;
 	virtual void clean() = 0;
 };
