+++ ServoSetup/protocoladapter.h
@@ -6,7 +6,7 @@
 
 class ProtocolAdapter : public SerialPort {
 public:
-	ProtocolAdapter(SerialPort *port, uint8_t addr, uint8_t id, uint8_t *buffer, int size);
+	ProtocolAdapter(SerialPort *port, uint8_t addr, uint8_t id, uint8_t *buffer, size_t size);
 	virtual ~ProtocolAdapter();
 	
 	virtual bool lock() {
@@ -16,8 +16,8 @@ public:
 	virtual void unlock() {
 	}
 
-	virtual uint16_t read(void *buf, uint16_t size);
-	virtual void write(const void *buf, uint16_t size);
+	virtual size_t read(void *buf, size_t size);
+	virtual void write(const void *buf, size_t size);
 	
 	virtual void clean() {
 	}
@@ -27,9 +27,9 @@ private:
 	const uint8_t id_;
 	const uint8_t addr_;
 	uint8_t *buffer_;
-	const int buffer_size_;
+	const size_t buffer_size_;
 	bool empty_;	/* true if no packet in buffer */
-	int size_;	/* received packet size */
+	size_t size_;	/* received packet size */
 	uint8_t *read_;	/* read pointer */
 };
